cmd,hack,internal,test: split off built-in scorecard plugins
