+++ hack/tests/scorecard-subcommand.sh
@@ -1,31 +1,27 @@
 #!/usr/bin/env bash
 
+source hack/lib/test_lib.sh
+
 DEST_IMAGE="quay.io/example/scorecard-proxy"
-CSV_PATH="deploy/olm-catalog/memcached-operator/0.0.3/memcached-operator.v0.0.3.clusterserviceversion.yaml"
 CONFIG_PATH=".test-osdk-scorecard.yaml"
 
 set -ex
 
-# build scorecard-proxy image (and delete intermediate builder image)
+# build scorecard-proxy image
 ./hack/image/build-scorecard-proxy-image.sh "$DEST_IMAGE"
+
+go build -o test/test-framework/scorecard/assets/osdk-scorecard-basic cmd/scorecard-basic/main.go
+trap_add 'rm test/test-framework/scorecard/assets/osdk-scorecard-basic' EXIT
+go build -o test/test-framework/scorecard/assets/osdk-scorecard-olm cmd/scorecard-olm/main.go
+trap_add 'rm test/test-framework/scorecard/assets/osdk-scorecard-olm' EXIT
 
 # the test framework directory has all the manifests needed to run the cluster
 pushd test/test-framework
-commandoutput="$(operator-sdk scorecard \
-  --cr-manifest deploy/crds/cache_v1alpha1_memcached_cr.yaml \
-  --cr-manifest deploy/crds/cache_v1alpha1_memcachedrs_cr.yaml \
-  --init-timeout 60 \
-  --csv-path "$CSV_PATH" \
-  --verbose \
-  --proxy-image "$DEST_IMAGE" \
-  --proxy-pull-policy Never \
-  2>&1)"
+commandoutput="$(operator-sdk scorecard 2>&1)"
 echo $commandoutput | grep "Total Score: 82%"
 
 # test config file
-commandoutput2="$(operator-sdk scorecard \
-  --proxy-image "$DEST_IMAGE" \
-  --config "$CONFIG_PATH")"
+commandoutput2="$(operator-sdk scorecard --config "$CONFIG_PATH")"
 # check basic suite
 echo $commandoutput2 | grep '^.*"error": 0,[[:space:]]"pass": 3,[[:space:]]"partialPass": 0,[[:space:]]"fail": 0,[[:space:]]"totalTests": 3,[[:space:]]"totalScorePercent": 100,.*$'
 # check olm suite
