+++ internal/pkg/scorecard/helpers.go
@@ -88,7 +88,7 @@ func CalculateResult(tests []scapiv1alpha1.ScorecardTestResult) scapiv1alpha1.Sc
 }
 
 // TestSuitesToScorecardOutput takes an array of test suites and generates a v1alpha1 ScorecardOutput object with the
-// provided name, description, and log
+// provided suites and log
 func TestSuitesToScorecardOutput(suites []TestSuite, log string) scapiv1alpha1.ScorecardOutput {
 	test := scapiv1alpha1.ScorecardOutput{
 		TypeMeta: metav1.TypeMeta{
