+++ test/test-framework/.test-osdk-scorecard.yaml
@@ -1,9 +1,2 @@
-cr-manifest:
-  - "deploy/crds/cache_v1alpha1_memcached_cr.yaml"
-  - "deploy/crds/cache_v1alpha1_memcachedrs_cr.yaml"
-init-timeout: 60
-csv-path: "deploy/olm-catalog/memcached-operator/0.0.3/memcached-operator.v0.0.3.clusterserviceversion.yaml"
-proxy-image: "scorecard-proxy"
-proxy-pull-policy: "Never"
 verbose: true
 output: json
