+++ test/test-framework/scorecard/assets/deploy
@@ -1,0 +1,1 @@
+../../deploy
