+++ test/test-framework/scorecard/bin/run-basic.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+./assets/osdk-scorecard-basic --config assets/.test-osdk-scorecard.yaml
