+++ test/test-framework/scorecard/bin/run-olm.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+./assets/osdk-scorecard-olm --config assets/.test-osdk-scorecard.yaml
